Add AvionRepository to ProprioServiceImpl
This commit integrates the AvionRepository into the ProprioServiceImpl. It enables the service to handle many-to-many relationships between Avion and Proprio, associating multiple "Avion" with one "Proprio" and vice versa. It updates the constructor of the ProprioServiceImpl and adjusts how the Proprio objects are created and saved, making the service even more versatile.